CRM-5 refactor: Extract permission-related methods from User model to HasPermissions trait
Description:
Refactored the code to improve separation of concerns by moving permission-related methods from the User model to a dedicated HasPermissions trait. This change enhances code organization and reusability.

Changes:
- Created HasPermissions trait to encapsulate permission-related logic.
- Moved the following methods from the User model to the HasPermissions trait:
  - `permissions`
  - `givePermissionTo`
  - `hasPermissionTo`
  - `getPermissionCacheKey`
- Updated the User model to use the HasPermissions trait.

Testing:
- Ensured that all existing tests for user permissions continue to pass, verifying that the trait integration maintains the expected functionality.
- Verified that caching for user permissions works as intended, reducing database queries and improving performance.
